+++ duckscript_sdk/src/sdk/std/env/os_name/mod_test.rs
@@ -1,0 +1,13 @@
+use super::*;
+use crate::test;
+use crate::test::CommandValidation;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_valid() {
+    test::run_script_and_validate(vec![create("")], "out = os_name", CommandValidation::Ignore);
+}